pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Simple task manager with CRUD API using FastAPI"
authors = [
    {name = "kspohodyaev",email = "konstantinpohodyaev@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "sqlalchemy (==2.0.40)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "httpx (==0.27.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "pytest-cov (>=6.2.1,<7.0.0)"
]

[tool.poetry.dependencies]



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
